Create staging environtment (#3)
* remove false warning

* update gitignore

* add new AnsibleInventory library

* create staging environment

* udpate Rakefile so that it supports staging

* add rubocop_todo

* create ec2.ini, mostly wiht defaults

* ignore rspec_status

+++ Rakefile
@@ -8,9 +8,11 @@ require "net/ssh"
 require "tempfile"
 require "pathname"
 require "highline/import"
-require "ansible/inventory/yaml"
 require "vagrant/serverspec"
 require "vagrant/ssh/config"
+require "retries"
+$LOAD_PATH.unshift(Pathname.new(File.dirname(__FILE__)) + "ruby" + "lib")
+require "ansible_inventory"
 
 def exec_and_abort_if_fail(cmd)
   status = system cmd
@@ -26,11 +28,14 @@ def vagrant(args)
 end
 
 def ansible_environment
-  ENV["ANSIBLE_ENVIRONMENT"] ? ENV["ANSIBLE_ENVIRONMENT"] : "virtualbox"
+  known_environment = %w[virtualbox staging prod]
+  env = ENV["ANSIBLE_ENVIRONMENT"] ? ENV["ANSIBLE_ENVIRONMENT"] : "virtualbox"
+  raise "unknown environment `#{env}`" unless known_environment.include?(env)
+  env
 end
 
-def inventory_file
-  "inventories/#{ansible_environment}/#{ansible_environment}.yml"
+def inventory_path
+  "inventories/#{ansible_environment}"
 end
 
 def sudo_password
@@ -38,7 +43,7 @@ def sudo_password
 end
 
 def sudo_password_required?(user)
-  user != "root" && user != "vagrant"
+  user != "root" && user != "vagrant" && user != "ec2-user"
 end
 
 def configure_sudo_password_for(run_as_user)
@@ -47,19 +52,49 @@ def configure_sudo_password_for(run_as_user)
     !ENV.key?("SUDO_PASSWORD")
 end
 
-desc "vagrant up"
-task :up do
-  vagrant "up"
+def plan_path
+  "terraform/plans/#{ansible_environment}"
 end
 
-desc "destroy VMs by `vagrant destroy`"
+desc "launch VMs"
+task :up do
+  case ansible_environment
+  when "virtualbox"
+    vagrant "up --no-provision"
+  when "staging"
+    sh "terraform apply #{plan_path}"
+
+    # make sure the cache is up-to-date
+    sh "#{inventory_path}/ec2.py --refresh-cache"
+
+    # make sure all hosts are ready for ansible play
+    retry_opts = {
+      max_tries: 10, base_sleep_seconds: 10, max_sleep_seconds: 30
+    }
+    with_retries(retry_opts) do |_attempt_number|
+      sh "ansible -i #{inventory_path} --ssh-common-args '-o \"UserKnownHostsFile /dev/null\" -o \"StrictHostKeyChecking no\"' --user 'ec2-user' -m ping all"
+    end
+  end
+end
+
+desc "destroy VMs"
 task :clean do
-  vagrant "destroy -f"
+  case ansible_environment
+  when "virtualbox"
+    vagrant "destroy -f"
+  when "staging"
+    sh "terraform destroy -force #{plan_path}"
+  end
 end
 
 desc "vagrant provision"
 task :provision do
-  vagrant "provision"
+  case ansible_environment
+  when "virtualbox"
+    vagrant "provision"
+  when "staging"
+    sh "ansible-playbook -i #{inventory_path} --ssh-common-args '-o \"UserKnownHostsFile /dev/null\" -o \"StrictHostKeyChecking no\"' --user 'ec2-user' playbooks/site.yml"
+  end
 end
 
 desc "all all tests; serverspec, integration, and rubocop"
@@ -68,6 +103,7 @@ task test: [
   # like to merge the branch first. or, the branch history will mess up.
   "test:rubocop:all",
   :up,
+  :provision,
   "test:serverspec:all",
   "test:integration:all"
 ] do
@@ -77,31 +113,52 @@ end
 namespace :test do
   desc "Prepare"
   task :prepare do
-    vagrant "up"
-    vagrant "provision"
+    case ansible_environment
+    when "virtualbox"
+      vagrant "up"
+      vagrant "provision"
+    end
   end
 
   desc "Provision"
   task :provision do
-    vagrant "provision"
+    case ansible_environment
+    when "virtualbox"
+      vagrant "provision"
+    else
+      warn "unknown environment `#{ansible_environment}`"
+      exit 1
+    end
   end
 
   desc "Restart VMs"
   task :restart do
-    vagrant "reload --provision"
+    case ansible_environment
+    when "virtualbox"
+      vagrant "reload --provision"
+    else
+      warn "unknown environment `#{ansible_environment}`"
+      exit 1
+    end
   end
 
   desc "Clean"
   task :clean do
-    begin
-      vagrant "destroy -f"
-    ensure
-      sh "rm -f *.vdi"
+    case ansible_environment
+    when "virtualbox"
+      begin
+        vagrant "destroy -f"
+      ensure
+        sh "rm -f *.vdi"
+      end
+    else
+      warn "unknown environment `#{ansible_environment}`"
+      exit 1
     end
   end
 
   namespace "serverspec" do
-    inventory = Ansible::Inventory::YAML.new(inventory_file)
+    inventory = AnsibleInventory.new(inventory_path)
     desc "Run serverspec on all hosts"
     task "all" do
       inventory.all_groups.each do |g|
@@ -109,10 +166,15 @@ namespace :test do
         inventory.all_hosts_in(g).each do |h|
           # XXX pass SUDO_PASSWORD to serverspec if the user is required to
           # type password
-          run_as_user = Vagrant::SSH::Config.for(h)["User".downcase]
+          run_as_user = case ansible_environment
+                        when "virtualbox"
+                          Vagrant::SSH::Config.for(h)["User".downcase]
+                        when "staging"
+                          "ec2-user"
+                        end
           configure_sudo_password_for(run_as_user)
           puts "running serverspec for #{g} on #{h} as user `#{run_as_user}`"
-          Vagrant::Serverspec.new(inventory_file).run(group: g, hostname: h)
+          Vagrant::Serverspec.new(inventory_path).run(group: g, hostname: h)
         end
       end
     end
@@ -122,10 +184,15 @@ namespace :test do
       task g.to_sym do |_t|
         inventory.all_hosts_in(g).each do |h|
           # XXX !DRY
-          run_as_user = Vagrant::SSH::Config.for(h)["User".downcase]
+          run_as_user = case ansible_environment
+                        when "virtualbox"
+                          Vagrant::SSH::Config.for(h)["User".downcase]
+                        when "staging"
+                          "ec2-user"
+                        end
           configure_sudo_password_for(run_as_user)
           puts "running serverspec for #{g} on #{h} as user `#{run_as_user}`"
-          Vagrant::Serverspec.new(inventory_file).run(group: g, hostname: h)
+          Vagrant::Serverspec.new(inventory_path).run(group: g, hostname: h)
         end
       end
     end
@@ -134,7 +201,7 @@ namespace :test do
   # XXX replace `serverspec` namespace with this when it is confirmed that the
   # tasks are acutually faster.
   namespace "para" do
-    inventory = Ansible::Inventory::YAML.new(inventory_file)
+    inventory = AnsibleInventory.new(inventory_path)
     inventory.all_groups.each do |g|
       next unless Dir.exist?("spec/serverspec/#{g}")
       inventory.all_hosts_in(g).each do |h|
@@ -143,7 +210,7 @@ namespace :test do
         #
         # desc "Run serverspec for group `#{g}` on #{h}"
         task "#{g}:#{h}" do |_t|
-          Vagrant::Serverspec.new(inventory_file)
+          Vagrant::Serverspec.new(inventory_path)
                              .run_with_fork(group: g, hostname: h)
         end
       end
@@ -167,6 +234,8 @@ namespace :test do
     run_as_user = case ansible_environment
                   when "virtualbox"
                     "vagrant"
+                  when "staging"
+                    "ec2-user"
                   when "prod"
                     ENV["USER"]
                   end
